todo ok

+++ app/src/main/java/com/example/myapplicationcanva/ConsumoData.kt
@@ -2,6 +2,6 @@ package com.example.myapplicationcanva
 
 data class ConsumoData(
     val periodo: String = "",
-    val consumo: Int = 0
+    val consumo: Float = 0f
 ) {
 }
